+ converted floats to doubles + Program outputs errors, a bug exists

+++ LBM-C/d2q9_boundary.cuh
@@ -2,7 +2,7 @@
 #define D2Q9_BOUNDARY_H
 
 // BOUNDARY CONDITION DEVICE FUNCTION PROTOTYPES
-__device__ Node zh_pressure_x(Node input, float rho_boundary);
-__device__ Node zh_pressure_X(Node input, float rho_boundary);
+__device__ Node zh_pressure_x(Node input, double rho_boundary);
+__device__ Node zh_pressure_X(Node input, double rho_boundary);
 
 #endif
